requirements.txt
Django
pip==25.1.1
djangorestframework==3.14.0
mysqlclient==2.1.1
Jinja2==3.1.2
python-dotenv==1.0.0
boto3==1.28.10
requests
djangorestframework-simplejwt
pyjwt
pillow
cryptography
httpx[http2]
# uwsgi
jwcrypto
django-db-connection-pool[mysql]
django-storages
django-cors-headers

# python3 -m venv venv
# .\venv\Scripts\activate
#  pip install -r .\requirements.txt
# django-admin startproject mini_issue_track_be .
# python3 manage.py startapp issues
